+++ django_extensions/management/modelviz.py
@@ -21,10 +21,17 @@ from django.utils.safestring import mark_safe
 from django.utils.translation import activate as activate_language
 
 try:
+    from django.utils.encoding import force_bytes
+except ImportError:
+    from django.utils.encoding import smart_str as force_bytes
+
+try:
     from django.db.models.fields.generic import GenericRelation
     assert GenericRelation
 except ImportError:
     from django.contrib.contenttypes.generic import GenericRelation
+
+from django_extensions.compat import get_apps
 
 
 __version__ = "1.0"
@@ -40,15 +47,18 @@ __contributors__ = [
     "Alexander Houben <alexander@houben.ch>",
     "Joern Hees <gitdev@joernhees.de>",
     "Kevin Cherepski <cherepski@gmail.com>",
+    "Jose Tomas Tocino <theom3ga@gmail.com>"
 ]
 
 
 def parse_file_or_list(arg):
     if not arg:
         return []
+    if isinstance(arg, (list, tuple, set)):
+        return arg
     if ',' not in arg and os.path.isfile(arg):
         return [e.strip() for e in open(arg).readlines()]
-    return arg.split(',')
+    return [e.strip() for e in arg.split(',')]
 
 
 def generate_dot(app_labels, **kwargs):
@@ -78,7 +88,7 @@ def generate_dot(app_labels, **kwargs):
 
     apps = []
     if all_applications:
-        apps = models.get_apps()
+        apps = get_apps()
 
     for app_label in app_labels:
         app = models.get_app(app_label)
@@ -134,14 +144,14 @@ def generate_dot(app_labels, **kwargs):
                 continue
 
             if verbose_names and appmodel._meta.verbose_name:
-                model['label'] = appmodel._meta.verbose_name.decode("utf8")
+                model['label'] = force_bytes(appmodel._meta.verbose_name)
             else:
                 model['label'] = model['name']
 
             # model attributes
             def add_attributes(field):
                 if verbose_names and field.verbose_name:
-                    label = field.verbose_name.decode("utf8")
+                    label = force_bytes(field.verbose_name)
                     if label.islower():
                         label = label.capitalize()
                 else:
@@ -183,7 +193,7 @@ def generate_dot(app_labels, **kwargs):
                 model['fields'] = sorted(model['fields'], key=lambda field: (not field['primary_key'], not field['relation'], field['label']))
 
             # FIXME: actually many_to_many fields aren't saved in this model's db table, so why should we add an attribute-line for them in the resulting graph?
-            #if appmodel._meta.many_to_many:
+            # if appmodel._meta.many_to_many:
             #    for field in appmodel._meta.many_to_many:
             #        if skip_field(field):
             #            continue
@@ -192,7 +202,7 @@ def generate_dot(app_labels, **kwargs):
             # relations
             def add_relation(field, extras=""):
                 if verbose_names and field.verbose_name:
-                    label = field.verbose_name.decode("utf8")
+                    label = force_bytes(field.verbose_name)
                     if label.islower():
                         label = label.capitalize()
                 else: