+++ django_extensions/management/technical_response.py
@@ -1,5 +1,5 @@
 import six
 
 
-def null_technical_500_response(request, exc_type, exc_value, tb):
+def null_technical_500_response(request, exc_type, exc_value, tb, status_code=500):
     six.reraise(exc_type, exc_value, tb)