+++ tests/test_clean_pyc.py
@@ -6,11 +6,10 @@ import six
 from django.core.management import call_command
 from django.test import TestCase
 
-from django_extensions.management.utils import get_project_root
-
 
 class CleanPycTests(TestCase):
     def setUp(self):
+        self.project_root = os.path.join('tests', 'testapp')
         self._settings = os.environ.get('DJANGO_SETTINGS_MODULE')
         os.environ['DJANGO_SETTINGS_MODULE'] = 'django_extensions.settings'
 
@@ -26,13 +25,13 @@ class CleanPycTests(TestCase):
         return pyc_glob
 
     def test_removes_pyc_files(self):
-        with self.settings(BASE_DIR=get_project_root()):
+        with self.settings(BASE_DIR=self.project_root):
             call_command('compile_pyc')
-        pyc_glob = self._find_pyc(get_project_root())
+        pyc_glob = self._find_pyc(self.project_root)
         self.assertTrue(len(pyc_glob) > 0)
-        with self.settings(BASE_DIR=get_project_root()):
+        with self.settings(BASE_DIR=self.project_root):
             call_command('clean_pyc')
-        pyc_glob = self._find_pyc(get_project_root())
+        pyc_glob = self._find_pyc(self.project_root)
         self.assertEqual(len(pyc_glob), 0)
 
     def test_takes_path(self):